test: extract UtilityTestSupport (Hotkey/URL/Screen-permission mocks) from TestSupport (#556, #434)
Sixth slice of splitting the large TestSupport.swift into domain files.
This slice groups three small OS-integration utility mocks —
MockHotkeyManager (HotkeyManaging), MockURLHandler (URLOpening), and
@MainActor MockScreenCapturePermissionAccess
(ScreenCapturePermissionAccessing) — into a shared UtilityTestSupport.swift.

Per codex's spec review, one file for these three is better than three
tiny files: each fakes a piece of the OS/system-integration surface,
and grouping them gives a shared "OS-integration mocks" home without
churn. They are non-adjacent in TestSupport.swift (lines 5-16, 300-314,
316-335 post-434e), so they are extracted from three separate ranges
and moved verbatim into the new file in original file order.

@MainActor is preserved on the line immediately above
MockScreenCapturePermissionAccess in the new file. The three
byte-diffs against TestSupport.swift@origin/main all return silent
(exit 0) — byte-identity proven for each block.

New file imports are the minimum required: Foundation (for URL) and
@testable import BugNarrator (for the protocols). No XCTest (unused),
no AppKit or SwiftUI (@MainActor does not require them).

Every other mock (both actor mocks, MockURLProtocol, StubExportReceiptStore)
and AppStateHarness stay in TestSupport.swift. Pure test-support move:
no product code, no helper renames, no behavior change. The full
BugNarratorTests suite passes at 662 tests (unchanged); docs-sync
validation passes.

Pre-build spec reviewed by codex in a fresh session (3 findings folded:
exact 662 baseline named against 0633a5d, single-file grouping
rationale strengthened, exact-ranges + byte-diff acceptance).
